Don't call conversion_proc for older Xservers
There was a bug in the Xserver for a while when it didn't call a driver's
conversion_proc to convert device space coordinates to screen space. To fix
this, vmmouse starting calling that method (since January), but if this is done
on older versions of X, the conversion gets done twice.

The fix is to add a configure check for the right versions of X and call
VMMouseConvertProc() only if we are compiling against an X that has the bug.

There's some whitespace cleanup for vmmouse.c too.