Add InputClass support.
The InputClass can match the tag "vmmouse". A sample xorg.conf.d snippet is
provided to try and make sure the driver will be loaded automatically in a
compatible VM also if the X server uses udev instead of HAL.

The configure script tries to determine if and where it makes sense to install
the udev rule and xorg.conf.d snippet but allows overriding both.

Thanks to Timo Aaltonen for the udev rules file and Dan Nicholson and Simon
Thum for review and suggestions for improvement.