Revert "Adjust the kernel name in the udev file."
This reverts commit 189dc266a66485d48303acbc1a4fbe4642c75e9f.

It seems that for some reason the device routing is breaking with this.
Possibly related to Win7 x64 hosts because it's not reproducible on
GNU/Linux hosts. Semantially we'd still like to match only on mouse
kernel but for the time being this seems to cause more problems
than it solves. Hence the revert.